rename app folder to `server`
